Removed absolute positioning for page-wrapper
Previously, the absolute positioning was used so that the pages would
occupy the same space while fading in/out. However, this caused the
<main> tag to not be aware of the height of the page, and thus stopped
the footer from getting pushed past the end of the viewport (it would
sit at the bottom of the viewport even if the content extended past).

To fix this, I set a timeout on the "setShow" of the PageWrapper2 class
when the page is to be faded in; in this way, the content of the new
page is not rendered until the old page has completely unmounted,
mitigating the positioning issue entirely.